Move interactive filters tests to their own file
